+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Develop
 
 Develop
 -----------------
+* [ENHANCEMENT] Allow  database store backend to support specification of schema in credentials file
 
 
 0.13.0